add add_dist_record method to UninstallPathSet
It moves the uninstall logic away from pip.wheel
